refactor: move Grid component out of Game
