+++ src/components/Game.tsx
@@ -1,6 +1,5 @@
 // Vendor
 import * as React from "react";
-import { Col, Grid, Row } from "react-native-easy-grid";
 import { Container } from "native-base";
 import { StyleSheet, View } from "react-native";
 import { Text } from "react-native";
@@ -8,9 +7,10 @@ import { Text } from "react-native";
 // Internal
 import { findWordConnections } from "../helpers/strings";
 import { getWordScore } from "../helpers/score";
+import { Grid } from "./Grid";
 import { Word } from "../types/interfaces";
 import { WordConnection } from "../types/enum";
-import { GAME_ROWS, SPECTATOR_INTERVAL } from "../config/settings";
+import { SPECTATOR_INTERVAL } from "../config/settings";
 
 export interface GameProps {
   allWords: Word[];
@@ -97,45 +97,10 @@ const Game: React.FC<GameProps> = (props) => {
     return () => clearInterval(intervalId);
   }, []); // Empty array leads to same behavior as `componentDidMount` (if it was a class)
 
-  const renderRow = (indexRow: number, last?: boolean) => {
-    const renderColumn = (letter: string, indexCol: number) => {
-      const innerComponent = last ? (
-        <View style={styles.viewText}>
-          <Text style={styles.text}>{letter}</Text>
-        </View>
-      ) : null;
-
-      return (
-        <Col key={`col-${indexCol}`} style={styles.cell}>
-          {innerComponent}
-        </Col>
-      );
-    };
-
-    const renderColumns = () => {
-      const word = words[words.length - 1];
-      const letters = word.englishWord.split("");
-
-      const columns = letters.map((v, i) => renderColumn(v, i));
-      return columns;
-    };
-
-    return <Row key={`row-${indexRow}`}>{renderColumns()}</Row>;
-  };
-
-  const renderRows = () => {
-    const lastIndexRow = GAME_ROWS - 1;
-    const allButLastRow = Array.from({ length: lastIndexRow }, (_v, i) =>
-      renderRow(i)
-    );
-
-    return [...allButLastRow, renderRow(lastIndexRow, true)];
-  };
-
   return (
     <Container>
       <Text style={styles.title}>Game</Text>
-      <Grid style={styles.grid}>{renderRows()}</Grid>
+      <Grid word={words[words.length - 1]} />
       <View style={styles.example}>
         <Text>{`Words: ${allWords.length}`}</Text>
         <Text>{`History: ${words.map((w) => w.englishWord)}`}</Text>
@@ -152,25 +117,12 @@ const Game: React.FC<GameProps> = (props) => {
 };
 
 const styles = StyleSheet.create({
-  cell: {
-    backgroundColor: "grey",
-    margin: 2,
-  },
   example: {
     backgroundColor: "yellow",
     flex: 1,
   },
-  grid: {
-    flex: 3,
-  },
-  text: { fontSize: 30 },
   title: {
     textAlign: "center",
-  },
-  viewText: {
-    alignItems: "center",
-    flex: 1,
-    justifyContent: "center",
   },
 });
